backend/requirements.txt
fastapi
python-multipart
gunicorn
uvicorn[standard]
python-pptx
groq
python-dotenv
slowapi
httpx
pypdf
python-docx
Pillow
motor
PyJWT
passlib[bcrypt]
openai>=1.0.0
fpdf2
pydantic-settings
